feat: sketch out the homepage copy
This diff starts sketching out the homepage copy. I used already
available material to patch this up. The intention is to have
a sense of what information is in the home page, so that I can
then move forward with styling the website. We will revisit this
content later on, ahead of publishing the website.