feat: sketch out the homepage copy (#16)
This diff starts sketching out the homepage copy. I used
already-available material to patch this up. The intention is to have a
sense of what information is in the home page, so that I can then move
forward with styling the website. We will revisit this content later on,
ahead of publishing the website.

Roughly speaking, this commit covers the first half of the content that
should be part of the homepage. Several sections have not been added
yet, and we only have empty placeholders representing them.

While there, take stock of the fact that running the website locally is
very helpful when making complex changes, and add instructions and
configuration for doing this by running `docker-compose`.