feat: create CNAME file (#21)
This tells GitHub that we expect this website to be server under the
`most.github.io.` domain. It should configure the `Host: ` header and
Let's Encrypt properly, after a little delay.

On our side, instead, we need to update the DNS to have `most.fbk.eu.`
be a `CNAME` for `fbk-most.github.io.`.