feat: use bootstrap 5.3.5 (#23)
This commit refactors the website to use bootstrap 5.3.5.

We changed the style a bit, but it needs to change anyway soon,
therefore, it is more or less fine(TM).

The advantage of integrating bootstrap is that we're now able to use
standard styles and additional features, such as the cookies warning.

Trying to keep things simple, I have vendored bootstrap in such a way
that the we don't need `npm` to install dependencies. It is a trade-off
between doing fancy things and keeping complexity at bay. We'll see how
well this serves us, and possibly do differently, if needed.